website/requirements.txt
streamlit>=1.42.0
Authlib>=1.3.2
openai>=0.27.8
chromadb
ollama
faiss-cpu